pyproject.toml
[tool.poetry]
name = "apitest"
version = "0.1.0"
description = "api test project"
authors = ["Lei Su <lei.susl@shopee.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
allure-pytest = "^2.12.0"
pymysql = "^1.0.2"
pyyaml = "^6.0"
requests = "^2.28.1"
pexpect = "^4.8.0"
pytest-rerunfailures = "^12.0"
lxml = "^4.9.3"
gentccode = "^0.2.12"
jsonschema = "^4.19.0"
pytest-dependency = "^0.6.0"
cryptography = "^42.0.6"
deepdiff = "^7.0.1"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
